Finalizando as Task: BD-22,BD-18, BD-20,BD-21, BD-23; transformando o models e service de 'Pessoa' Assicrona e adicionando teste unitarios

+++ Banco/WebAbil8-Sistema_Verificação_dupla.slnx/Services/ITentativaAcessoRepository.cs
@@ -5,10 +5,10 @@ namespace WebAbil8_Sistema_Verificação_dupla.slnx.Services
     public interface ITentativaAcessoRepository
     {
         TentativaAcesso Adicionar(TentativaAcesso tentativa);
-        TentativaAcesso BuscarPorId(long id);
+        TentativaAcesso BuscarPorId(int id);
         List<TentativaAcesso> ListarTodos();
         TentativaAcesso Atualizar(TentativaAcesso tentativa);
-        void Remover(long id);
+        void Remover(int id);
         TentativaAcesso Registrar(TentativaAcesso tentativa);
         List<TentativaAcesso> ListarComFiltros(long? pessoaId, int? ambienteId, bool? acessoLiberado, DateTime? dataInicio, DateTime? dataFim);
         List<TentativaAcesso> ListarPorPessoa(long pessoaId);
